requirements.txt
# Cloud Cost Knowledge Graph - Dependencies
# Pinned to versions already installed on this system

# Core Backend
fastapi==0.128.0
uvicorn==0.40.0
neo4j==6.1.0
pydantic==2.12.5
python-dotenv==1.0.1
requests==2.32.3

# UI
streamlit==1.53.0

# ML / Embeddings
sentence-transformers==5.2.0
torch==2.10.0
scikit-learn==1.6.1
numpy==2.4.1
pandas==2.3.3

# LLM Support (Gemini called via REST API — no SDK needed)
openai==2.24.0

# Data Loading
openpyxl>=3.1.0
xlrd>=2.0.1

# Graph Visualization
pyvis>=0.3.2

# API
httpx==0.28.1